requirements.txt
# FastAPI and server dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.21
asyncpg>=0.28.0
alembic>=1.12.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client for external API calls
httpx>=0.25.0

# Image processing and OCR
Pillow>=10.0.1
pytesseract>=0.3.10
pdf2image>=1.16.3
opencv-python>=4.8.1.78

# Utilities
numpy>=1.26.0
pandas>=2.1.1
python-dateutil>=2.8.2

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4